refactor(frontend): split YearInReviewPage into year-in-review/ module
