+++ frontend/src/App.tsx
@@ -34,7 +34,7 @@ const pageImports = {
   TrendsForecastsPage: () => import('@/pages/TrendsForecastsPage'),
   ComparisonPage: () => import('@/pages/ComparisonPage'),
   BudgetPage: () => import('@/pages/BudgetPage'),
-  YearInReviewPage: () => import('@/pages/YearInReviewPage'),
+  YearInReviewPage: () => import('@/pages/year-in-review/YearInReviewPage'),
   SettingsPage: () => import('@/pages/SettingsPage'),
   AnomalyReviewPage: () => import('@/pages/AnomalyReviewPage'),
   GoalsPage: () => import('@/pages/GoalsPage'),
